Moving explicit and implicit micro-experiments from exp01_gem to exp04_update_of_c
